pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-research"
version = "0.1.0"
description = "OSINT Research Automation System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Async HTTP
    "httpx>=0.27.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Redis
    "redis[hiredis]>=5.0.0",

    # Vector DB
    "qdrant-client>=1.9.0",

    # Task Queue
    "celery[redis]>=5.3.0",

    # LLM & Embeddings
    "openai>=1.12.0",
    "sentence-transformers>=2.5.0",

    # NLP
    "rapidfuzz>=3.6.0",

    # Retry & resilience
    "tenacity>=8.2.0",

    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.7",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.1.0",

    # Document text extraction
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",

    # GitHub repo ingestion
    "gitingest",

    # BibTeX parsing
    "bibtexparser>=1.4.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
